WIP: data collection process
Every webpage has it's own google tracking tag, but that doesn't matter.
It only reads the contents inside <body><\body>. I finally got it to
work such that it can read a story title, tags, and present it in the
main list.

The next steps is to adjust it so the story page doesn't have the tags.